new options: checkout (true), archive(false)

+++ checkout.sh
@@ -2,36 +2,42 @@
 set -o errexit -o nounset -o pipefail
 cd "`dirname $0`"
 
-if [ ! -d "depot_tools" ]; then
-  git clone 'https://chromium.googlesource.com/chromium/tools/depot_tools.git'
-fi
-export PATH="${PWD}/depot_tools:${PATH}"
+if [ "${checkout:-true}" != "false" ]; then
 
-skia_branch=${skia_branch:-chrome/${version}}
-
-if [ -d "skia" ]; then
-  pushd skia
-  if [ -n "$(git branch --list ${skia_branch})" ]; then
-    echo "> Advancing ${skia_branch}"
-    git checkout -B ${skia_branch}
-    git fetch
-    git reset --hard origin/${skia_branch}
-  else
-    echo "> Fetching ${skia_branch}"
-    git fetch origin ${skia_branch}:remotes/origin/${skia_branch}
-    git checkout -b ${skia_branch}
+  if [ ! -d "depot_tools" ]; then
+    git clone 'https://chromium.googlesource.com/chromium/tools/depot_tools.git'
   fi
-else
-  echo "> Cloning ${skia_branch}"
-  time git clone https://skia.googlesource.com/skia --quiet --branch ${skia_branch}
-  pushd skia
-fi
+  export PATH="${PWD}/depot_tools:${PATH}"
 
-if [ -n "${skia_commit:-}" ]; then
-  echo "> Checking out ${skia_commit}"
-  git -c advice.detachedHead=false checkout ${skia_commit}
+  skia_branch=${skia_branch:-chrome/${version}}
+
+  if [ -d "skia" ]; then
+    pushd skia > /dev/null
+    if [ -n "$(git branch --list ${skia_branch})" ]; then
+      echo "> Advancing ${skia_branch}"
+      git checkout -B ${skia_branch}
+      git fetch
+      git reset --hard origin/${skia_branch}
+    else
+      echo "> Fetching ${skia_branch}"
+      git fetch origin ${skia_branch}:remotes/origin/${skia_branch}
+      git checkout -b ${skia_branch}
+    fi
+  else
+    echo "> Cloning ${skia_branch}"
+    time git clone https://skia.googlesource.com/skia --quiet --branch ${skia_branch}
+    pushd skia > /dev/null
+  fi
+
+  if [ -n "${skia_commit:-}" ]; then
+    echo "> Checking out ${skia_commit}"
+    git -c advice.detachedHead=false checkout ${skia_commit}
+  else
+    echo "> Using $(git rev-parse HEAD)"
+  fi
+
 else
-  echo "> Using $(git rev-parse HEAD)"
+  pushd skia > /dev/null
 fi
 
 if [ -z "${release:-}" ]; then
